rg-17: * decompose types to files
