+++ frontend/src/bundles/common/components/calendar/constants/calendar.constants.ts
@@ -1,10 +1,11 @@
-import { type DateDto } from '~/bundles/common/types/calendar/calendar-date.dto';
+
+import { type CalendarDate } from '~/bundles/common/types/calendar/calendar-date';
 import { type CalendarMonth } from '~/bundles/common/types/types';
 
 const yearRegex = /\d{4}/g;
 const monthRegex = /[A-Za-z]+/;
 
-const DEFAULT_DATE_PAYLOAD: DateDto = {
+const DEFAULT_DATE_PAYLOAD: CalendarDate = {
     present: false,
     month: null,
     year: new Date().getFullYear()
