--- frontend/src/bundles/common/types/calendar/calendar-date.dto.ts
@@ -1,8 +1,0 @@
-import { type CalendarMonth } from './calendar.types';
-
-type DateDto = {
-    present: boolean;
-    month: CalendarMonth | null,
-    year: number
-};
-export { type DateDto };
