+++ frontend/src/bundles/common/types/calendar/calendar-date.ts
@@ -1,0 +1,9 @@
+import { type CalendarMonth } from './calendar-month';
+
+type CalendarDate = {
+    present: boolean;
+    month: CalendarMonth | null;
+    year: number | null;
+};
+
+export { type CalendarDate };
