+++ frontend/src/bundles/common/types/calendar/calendar-month.ts
@@ -1,0 +1,6 @@
+type CalendarMonth = {
+    name: string;
+    num: number;
+};
+
+export { type CalendarMonth };
