--- frontend/src/bundles/common/types/calendar/calendar.types.ts
@@ -1,12 +1,0 @@
-type CalendarDate = {
-    present: boolean;
-    month: CalendarMonth | null;
-    year: number | null;
-};
-
-type CalendarMonth = {
-    name: string;
-    num: number;
-};
-
-export { type CalendarDate, type CalendarMonth };
