+++ frontend/src/bundles/common/types/types.ts
@@ -1,6 +1,6 @@
 export { type AsyncThunkConfig } from './async-thunk-config.type.js';
-export { type CalendarDate } from './calendar/calendar.types.js';
-export { type CalendarMonth } from './calendar/calendar.types.js';
+export { type CalendarDate } from './calendar/calendar-date.js';
+export { type CalendarMonth } from './calendar/calendar-month.js';
 export {
     type ServerErrorDetail,
     type ServerErrorResponse,
